Broadcast to a single peer
The problem with blasting transactions to multiple peers is that it
increases the chance of landing on a surveillance node. This isn't a
problem when using Tor; it is mainly a problem for clearnet connections.

This change ensures that we broadcast to just one peer and then wait for
a roundtrip confirmation in the form of an `inventory` message from
other connected peers.

This is marginally slower than the blasting strategy but it increases
both privacy for clearnet connections and gives us certainty in the form
of verifying that our transaction has propagated through the network.

The selected broadcast peer is given 10 seconds for our transaction(s)
to appear on the network, after which we rotate the peer.

Other changes:

- bump version to 0.4.0 for both lib and bin
- redo the `Report` struct to better reflect completion outcomes
- deprecate the `--tesnet` switch in bin in favor of `--network`
- remove the `send_unsolicited` option (unnecessary)